+++ src/json-data/json-data.js
@@ -5,6 +5,8 @@ import { Albums } from "./albums.js";
 import { Songs } from "./songs.js";
 import { Players } from "./players.js";
 import { Recipes } from "./recipes.js";
+import { Students } from "./student-scores.js";
+import { Countries } from "./countries.js";
 
 export class JSONData {
   constructor() {}
@@ -15,4 +17,6 @@ export class JSONData {
   static Songs = Songs ? Songs : [];
   static Players = Players ? Players : [];
   static Recipes = Recipes ? Recipes : [];
+  static Students = Students ? Students : [];
+  static Countries = Countries ? Countries : [];
 }